Remove forceRender check from html-client-preview
The forceRender being true means that any time the story is returning a node (instead of a string) it's not re-rendering the component. This is problematic when you are affecting changes to the node that goes beyond html elements (where a workaround like node.outerHTML would suffice).

I think it's safer to re-render the node then to assume the dom never changes, otherwise addons like knobs become useless. I'm open to discussing how the forceRender check could be kept such as adding an escape hatch to allow specific stories to specify to re-render on change.